Separates ACL and PAC update
